% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateFormatRoutines.R
\name{as.calWeek}
\alias{as.calWeek}
\title{Return the calender week of the date (ISO 8601)}
\usage{
as.calWeek(date)
}
\arguments{
\item{date}{to be transformed}
}
\value{
the numeric calendar week
}
\description{
Return the calender week of the date (ISO 8601)
}
