% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateFormatRoutines.R
\name{as.days_since_d0}
\alias{as.days_since_d0}
\title{Return for each date the days since the minimum date in the vector.}
\usage{
as.days_since_d0(dates)
}
\arguments{
\item{dates}{a vector of dates}
}
\value{
a vector of days (integers)
}
\description{
Return for each date the days since the minimum date in the vector.
}
