% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateFormatRoutines.R
\name{as.days_since_global_d0}
\alias{as.days_since_global_d0}
\title{Return for each date in the vector the days since the given date "minDate".}
\usage{
as.days_since_global_d0(dates, minDate)
}
\arguments{
\item{dates}{a vector of dates}

\item{minDate}{a date}
}
\value{
a vector of days (integers)
}
\description{
Return for each date in the vector the days since the given date "minDate".
}
