% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateFormatRoutines.R
\name{as.lastDayOfWeek}
\alias{as.lastDayOfWeek}
\title{Return a representative date of the calender week. Here: the last day.}
\usage{
as.lastDayOfWeek(week, year)
}
\arguments{
\item{week}{a calendar week (integer)}

\item{year}{a year (integer) in the format yyyy}
}
\value{
the date of the last day of the calendar week
}
\description{
Return a representative date of the calender week. Here: the last day.
}
