% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateFormatRoutines.R
\name{asDayOfYear}
\alias{asDayOfYear}
\title{Return the day of the year of the date as "yyyyddd".}
\usage{
asDayOfYear(date)
}
\arguments{
\item{date}{to be transformed}
}
\value{
the numeric day of the year in the format "yyyyddd"
}
\description{
Return the day of the year of the date as "yyyyddd".
}
