% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateFormatRoutines.R
\name{days.as.Date}
\alias{days.as.Date}
\title{Return the dates when adding the given vector of integer days to the given date "minDate"}
\usage{
days.as.Date(days, minDate)
}
\arguments{
\item{days}{a vector of days (integers)}

\item{minDate}{a date}
}
\value{
a vector of dates
}
\description{
Return the dates when adding the given vector of integer days to the given date "minDate"
}
