% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateFormatRoutines.R
\name{daysAsDate}
\alias{daysAsDate}
\title{Return the dates when adding the given vector of integer days to the given date "minDate"}
\usage{
daysAsDate(days, minDate)
}
\arguments{
\item{days}{a vector of days (integers)}

\item{minDate}{a date}
}
\value{
a vector of dates
}
\description{
Return the dates when adding the given vector of integer days to the given date "minDate"
}
