% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateFormatRoutines.R
\name{doy.as.Date}
\alias{doy.as.Date}
\title{Return the date from the day of the year of the date as yyyyddd.}
\usage{
doy.as.Date(doy)
}
\arguments{
\item{doy}{day of the year in the format "yyyyddd"}
}
\value{
the date
}
\description{
Return the date from the day of the year of the date as yyyyddd.
}
