% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRoutines.R
\name{plot_Re}
\alias{plot_Re}
\title{Plot Re estimates.}
\usage{
plot_Re(re.table, outputFile = NULL, title = NULL)
}
\arguments{
\item{re.table}{a data frame containing a column "value" with R_e estimates, values "lower" and "upper" for the confidence interval and a column "date"}

\item{outputFile}{(optional) The Re plot is written to this file}

\item{title}{(optional) a vector of time points which should be interpolated}
}
\description{
Plotting the Re estimtes with its confidence interval and saving it at the given location.
}
