% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRoutines.R
\name{plot_smoothed_phi_with_new_cases}
\alias{plot_smoothed_phi_with_new_cases}
\title{Plot smoothed phi estimates over time.}
\usage{
plot_smoothed_phi_with_new_cases(
  smooth.table,
  point.table = NULL,
  cases.table = NULL,
  outputFile = NULL,
  outputFileDots = NULL,
  title = NULL
)
}
\arguments{
\item{smooth.table}{a data frame containing the smoothed Medians as output from smooth_point_estimates}

\item{point.table}{(optional) a data frame containing the point estimates for phi in the column "value", a "date", and the "sampleSize"}

\item{cases.table}{(optional) The data frame containing the reported cases with the rolling average in column "new_cases_avrg" and "date}

\item{outputFile}{(optional) The plot without point estimates is written to this file}

\item{outputFileDots}{(optional) The plot with point estimates is written to this file}

\item{title}{(optional) For the title of the plot}
}
\description{
Plotting the smoothed phi estimates, both with and without point estimates (if the output files are given).
If present, the reported cases are added to plot with the scale on the right y-axis.
}
