% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothingRoutines.R
\name{smooth_point_estimates}
\alias{smooth_point_estimates}
\title{Smoothing over of the dated point estimates}
\usage{
smooth_point_estimates(
  input.table,
  ts,
  weights = NULL,
  N_sampling = NULL,
  p_sample = NULL,
  width = NULL
)
}
\arguments{
\item{input.table}{a data frame containing a column "value" with the point estimates and a column "t" wiht the time points}

\item{ts}{a vector of time points which should be interpolated}

\item{weights}{(optional) a weight between 0 and 1 for each value in the table, for being sampled (default: all equal)}

\item{N_sampling}{(optional) number of resampling to draw the median and percentiles from (default: 1000)}

\item{p_sample}{(optional) proportion of point estimates to sample (default: 0.5)}

\item{width}{(optional) windowsize for smoothing (default: 7)}
}
\value{
a data frame containing the smoothed values for each time point  ("t" = time points in ts, "smoothMedian" = median, "smooth5" = 5th percentile and "smooth95" = 95th percentile)
}
\description{
The function resamples a certain proportion of the given point estimates, if given with the respective weight, if not randomly.
For each resampling the values are collected, if no data is available for certain days, the values are interpolated.
The median and 5th and 95th percentiles at each time point are smoothed with a 2-sided convolution filter of the given window size.
}
